core: replace with new api version
