+++ Blasphemous.BootsOfPleading.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blasphemous.BootsOfPleading", "Blasphemous.BootsOfPleading\Blasphemous.BootsOfPleading.csproj", "{4F89A77B-8E87-48AF-B02F-0D6FF56765CB}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4F89A77B-8E87-48AF-B02F-0D6FF56765CB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4F89A77B-8E87-48AF-B02F-0D6FF56765CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4F89A77B-8E87-48AF-B02F-0D6FF56765CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4F89A77B-8E87-48AF-B02F-0D6FF56765CB}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {45AE5E7C-7DFB-46DF-A750-31D4A54E927C}
+	EndGlobalSection
+EndGlobal
