--- BootsOfPleading.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.31605.320
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "BootsOfPleading", "BootsOfPleading\BootsOfPleading.csproj", "{40962854-D9A6-4874-BF10-7632A455DCF2}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{40962854-D9A6-4874-BF10-7632A455DCF2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40962854-D9A6-4874-BF10-7632A455DCF2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40962854-D9A6-4874-BF10-7632A455DCF2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40962854-D9A6-4874-BF10-7632A455DCF2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {AE0018A4-F084-45A0-BD76-01835FD5C0A9}
-	EndGlobalSection
-EndGlobal
