--- BootsOfPleading/CheckTrapDerived.cs
@@ -1,16 +1,0 @@
-﻿using Gameplay.GameControllers.Entities;
-using UnityEngine;
-
-namespace BootsOfPleading
-{
-    public class CheckTrapDerived : CheckTrap
-    {
-		private void OnTriggerExit2D(Collider2D other)
-		{
-			if (other.gameObject.CompareTag("SpikeTrap"))
-			{
-				Main.SpikeProtection.LeftSpikes();
-			}
-		}
-	}
-}
