--- BootsOfPleading/Main.cs
@@ -1,17 +1,0 @@
-﻿using BepInEx;
-
-namespace BootsOfPleading
-{
-    [BepInPlugin(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION)]
-    [BepInDependency("com.damocles.blasphemous.modding-api", "1.3.3")]
-    [BepInProcess("Blasphemous.exe")]
-    public class Main : BaseUnityPlugin
-    {
-        public static SpikeProtection SpikeProtection { get; private set; }
-
-        private void Start()
-        {
-            SpikeProtection = new SpikeProtection(PluginInfo.PLUGIN_GUID, PluginInfo.PLUGIN_NAME, PluginInfo.PLUGIN_VERSION);
-        }
-    }
-}
